% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\docType{class}
\name{CentipedeMixEm-class}
\alias{CentipedeMixEm-class}
\alias{CentipedeMixEm}
\title{Actual EM algorithm for CENTIPEDE}
\arguments{
\item{k}{number of mixed distributions}

\item{l}{number of rows in X}

\item{s}{number of columns in X}

\item{X}{observed experimental data matrix}

\item{R}{number of reads in each motif, sum of each row in X}

\item{G}{prior information matrix (PWM+Cons.Score+TSS)}

\item{Z}{binary variable used to initialize beta's}

\item{pi_mat}{prior probability matrix}

\item{B}{beta's for the logistic regression used to calculate pi_l}

\item{a0}{parameters for negative binomial distribution unbound}

\item{a1}{parameters for negative binomial distribution bound}

\item{tau0}{parameters for negative binomial distribution unbound}

\item{tau1}{parameters for negative binomial distribution bound}

\item{lambda}{parameters for the multinomial distribution}

\item{prob_mat}{posterior probability matrix}

\item{log_prob_vec}{log of posterior probability of bound}

\item{max_log}{maximum of log posterior bound probability}

\item{loglik}{log likelihood}

\item{tol}{threshold for testing convergence}

\item{logittau0}{logit scale of tau0}

\item{logittau1}{logit scale of tau1}

\item{logitbeta}{logit scale of beta}
}
\value{
A list of convergence, updated parameters, posterior probabilities, iterations, log likelihood list, bound or unbound status
#'@export
}
\description{
Actual EM algorithm for CENTIPEDE
}

