% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{compare_results}
\alias{compare_results}
\title{Calculate statistcs between EM results and ChIP-seq data}
\usage{
compare_results(results, region_labels, chip_file)
}
\arguments{
\item{results}{a binary vectory of bound status predictions}

\item{region_labels}{dataframe of coordinates corresponding to results with columns 'chr','start','end' and rows as the coordinates in "chr:start-end" format}

\item{chip_file}{a bed file of ChIP-seq peaks}
}
\value{
A list of statistics including precision, FDR, FPR, sensitivity,and a dataframe labeled_df which is results with added columns for predicted labels and ChIP-seq labels
}
\description{
Calculate statistcs between EM results and ChIP-seq data
}
