% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{plot_cutsites}
\alias{plot_cutsites}
\title{Plots the cutsite probabilities (lambda parameter) for EM results}
\usage{
plot_cutsites(lambda, name, out)
}
\arguments{
\item{lambda}{a vector of lambda probabilities derived from the multinomial distribution}

\item{name}{the name of the run for the title of the plot}

\item{out}{the output file to save plot to}
}
\description{
Plots the cutsite probabilities (lambda parameter) for EM results
}
\examples{
plot_cutsites(result$lambda,"my experiment","my_experiment_cutsites.png")
}
