% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{plot_roc}
\alias{plot_roc}
\title{Plots ROC curve in precrec's style}
\usage{
plot_roc(labelled_df)
}
\arguments{
\item{labelled_df}{a dataframe outputted by compare_results or with columns for bound_status (predicted labels) and label}
}
\description{
Plots ROC curve in precrec's style
}
