% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EM_algorithm.R
\name{run_centipede}
\alias{run_centipede}
\title{Run CENTIPEDE model}
\usage{
run_centipede(input_list, cons_file = NULL)
}
\arguments{
\item{input_list}{List containing PWMScan, TSS data, and X}

\item{cons_file}{Optional matrix including conservation scores}
}
\value{
The results from running the CENTIPEDE model (bound status)
}
\description{
Run CENTIPEDE model
}
